generateTest functio -> useGenerateHook file
