+++ src/components/TestCase/HooksTestCase.tsx
@@ -1,7 +1,7 @@
 import React, { useContext, useRef, useEffect } from 'react';
 import { DragDropContext, Droppable, DropResult } from 'react-beautiful-dnd';
 import { GlobalContext } from '../../context/reducers/globalReducer';
-import { createFile } from '../../context/actions/globalActions';
+import { createFile, setFilePath } from '../../context/actions/globalActions';
 import styles from './TestCase.module.scss';
 import { HooksTestCaseContext } from '../../context/reducers/hooksTestCaseReducer';
 import {
@@ -12,10 +12,7 @@ import HooksTestMenu from '../TestMenu/HooksTestMenu';
 import HooksTestStatements from './HooksTestStatements';
 import { HooksStatements } from '../../utils/hooksTypes';
 import HooksHelpModal from '../TestHelpModals/HooksHelpModal';
-
-const remote = window.require('electron').remote;
-const beautify = remote.require('js-beautify');
-const path = remote.require('path');
+import useGenerateTest from '../../context/useGenerateTest.jsx';
 
 const HooksTestCase = () => {
   const [{ hooksTestStatement, hooksStatements, modalOpen }, dispatchToHooksTestCase] = useContext(
@@ -58,142 +55,12 @@ const HooksTestCase = () => {
     );
     dispatchToHooksTestCase(updateStatementsOrder(reorderedStatements));
   };
-  let testFileCode = 'import React from "react";';
-  const generatHookFile = () => {
-    return (
-      addHooksImportStatements(),
-      addHooksTestStatements(),
-      (testFileCode = beautify(testFileCode, {
-        indent_size: 2,
-        space_in_empty_paren: true,
-        e4x: true,
-      }))
-    );
-  };
 
-  const addHooksImportStatements = () => {
-    hooksStatements.forEach((statement: any) => {
-      switch (statement.type) {
-        case 'hook-updates':
-          return addRenderHooksImportStatement(), createPathToHooks(statement);
-        case 'hookRender':
-          return addRenderHooksImportStatement(), createPathToHooks(statement);
-        case 'context':
-          return addContextImportStatements(), createPathToContext(statement);
-        default:
-          return statement;
-      }
-    });
-    testFileCode += '\n';
-  };
-
-  const addHooksTestStatements = () => {
-    testFileCode += `\n test('${hooksTestStatement}', () => {`;
-    hooksStatements.forEach((statement: any) => {
-      switch (statement.type) {
-        case 'hook-updates':
-          return addHookUpdates(statement);
-        case 'hookRender':
-          return addHookRender(statement);
-        case 'context':
-          return addContext(statement);
-        default:
-          return statement;
-      }
-    });
-    testFileCode += '});';
-    testFileCode += '\n';
-  };
-
-  // Hooks Import Statements
-  const addRenderHooksImportStatement = () => {
-    if (!testFileCode.includes(`import '@testing-library/jest-dom/extend-expect';`)) {
-      testFileCode += `import '@testing-library/jest-dom/extend-expect';`;
-    }
-    if (!testFileCode.includes(`import { renderHook, act } from '@testing-library/react-hooks';`)) {
-      testFileCode += `import { renderHook, act } from '@testing-library/react-hooks';`;
-    }
-  };
-
-  // Hooks Filepath
-  const createPathToHooks = (statement: any) => {
-    let filePath = path.relative(projectFilePath, statement.hookFilePath);
-    filePath = filePath.replace(/\\/g, '/');
-    testFileCode += `import ${statement.hook} from '../${filePath}';`;
-  };
-
-  // Context Import Statements
-  const addContextImportStatements = () => {
-    if (!testFileCode.includes(`import '@testing-library/jest-dom/extend-expect';`)) {
-      testFileCode += `import '@testing-library/jest-dom/extend-expect';`;
-    }
-    if (!testFileCode.includes(`import { render } from '@testing-library/react';`)) {
-      testFileCode += `import { render } from '@testing-library/react';`;
-    }
-  };
-
-  // Context Filepath
-  const createPathToContext = (statement: any) => {
-    let filePath = path.relative(projectFilePath, statement.contextFilePath);
-    filePath = filePath.replace(/\\/g, '/');
-    testFileCode += `import { ${statement.providerComponent}, ${statement.consumerComponent}, ${statement.context} } from '../${filePath}';`;
-  };
-
-  // Hook: Updates Jest Test Code
-  const addHookUpdates = (hookUpdates: any) => {
-    testFileCode += `const {result} = renderHook (() => ${hookUpdates.hook}());
-    act(() => {
-      result.current.${hookUpdates.callbackFunc}();
-    });
-    expect(result.current.${hookUpdates.managedState}).toBe(${hookUpdates.updatedState})`;
-  };
-
-  // Hook: Renders Jest Test Code
-  const addHookRender = (hookRender: any) => {
-    testFileCode += `const {result} = renderHook(() => ${hookRender.hook}(${hookRender.parameterOne}))
-    expect(result.current.${hookRender.returnValue}).toBe(${hookRender.expectedReturnValue})`;
-  };
-
-  // Context Jest Test Code
-  const addContext = (context: any) => {
-    if (context.queryValue === 'shows_default_value') {
-      testFileCode += `const mockValue = {Data: '${context.values}'}
-      const { ${context.querySelector} } = render(<${context.consumerComponent}/>)
-      expect(${context.querySelector}(mockValue.Data)).${context.queryVariant}('${context.values}')`;
-    }
-    if (context.queryValue === 'shows_value_from_provider') {
-      testFileCode += `const mockValue = {Data: '${context.values}'}
-      const { ${context.querySelector} } = render (
-        <${context.context}.Provider value={mockValue}>
-          <${context.consumerComponent}/>
-        </${context.context}.Provider>
-      )
-      expect(${context.querySelector}(mockValue.Data)).${context.queryVariant}('${context.values}')`;
-    }
-    if (context.queryValue === 'component_provides_context_value') {
-      testFileCode += `const mockValue = {Data: '${context.values}'}
-      const { ${context.querySelector} } = render (
-        <${context.providerComponent} value={mockValue}>
-          <${context.context}.Consumer>
-          {value => <span>Recieved: {value} </span>}
-          <${context.context}.Consumer/>
-        </${context.providerComponent}>
-      )
-      expect(${context.querySelector}(/^Recieved:/).textContent).${context.queryVariant}('${context.values}')`;
-    }
-    if (context.queryValue === 'renders_providers_+_consumers_normally') {
-      testFileCode += `const mockValue = {Data: '${context.values}'}
-      const { ${context.querySelector} } = render (
-        <${context.providerComponent} value={mockValue}>
-          <${context.consumerComponent}/>
-        </${context.providerComponent}>
-      )
-      expect(${context.querySelector}(mockValue.Data).textContent).${context.queryVariant}('${context.values}')`;
-    }
-  };
+  const generateTest = useGenerateTest('hooks', projectFilePath);
 
   const fileHandle = () => {
-    dispatchToGlobal(createFile(generatHookFile()));
+    dispatchToGlobal(createFile(generateTest({ hooksTestStatement, hooksStatements })));
+    dispatchToGlobal(setFilePath(''));
   };
 
   return (
