+++ src/components/TestCase/ReactTestCase.jsx
@@ -22,10 +22,7 @@ import {
   reactTestCaseState,
   reactTestCaseReducer,
 } from '../../context/reducers/reactTestCaseReducer';
-//
-const remote = window.require('electron').remote;
-const path = remote.require('path');
-const beautify = remote.require('js-beautify');
+import useGenerateTest from '../../context/useGenerateTest.jsx';
 
 const ReactTestCase = () => {
   //changes to pull down context
@@ -34,6 +31,7 @@ const ReactTestCase = () => {
     reactTestCaseState
   );
   //
+
   const { describeBlocks, itStatements, statements, modalOpen } = reactTestCase;
   const [{ mockData }, dispatchToMockData] = useContext(MockDataContext);
   const [{ filePathMap, projectFilePath }, dispatchToGlobal] = useContext(GlobalContext);
@@ -55,141 +53,11 @@ const ReactTestCase = () => {
     dispatchToReactTestCase(updateItStatementText(text, itId));
   };
 
-  let testFileCode = 'import React from "react";';
-
-  const generatReactFile = () => {
-    return (
-      addComponentImportStatement(),
-      addReactImportStatements(),
-      addMockData(),
-      addDescribeBlocks(),
-      (testFileCode = beautify(testFileCode, {
-        brace_style: 'collapse, preserve-inline',
-        indent_size: 2,
-        space_in_empty_paren: true,
-        e4x: true,
-      }))
-    );
-  };
-  const addMockData = () => {
-    mockData.forEach((mockDatum) => {
-      let fieldKeys = createMockDatumFieldKeys(mockDatum);
-      testFileCode += `const mock${
-        mockDatum.name.charAt(0).toUpperCase() + mockDatum.name.slice(1)
-      } = build('${mockDatum.name}').fields({ ${fieldKeys} })();`;
-    });
-    testFileCode += '\n';
-  };
-  // const handleAddMockData = () => {
-  //   dispatchToMockData(addMockData());
-  // };
-
-  const createMockDatumFieldKeys = (mockDatum) => {
-    return mockDatum.fieldKeys.reduce((fieldKeysCode, mockDatum) => {
-      return fieldKeysCode + `${mockDatum.fieldKey}: fake(f => f.random.${mockDatum.fieldType}()),`;
-    }, '');
-  };
-
-  // React Import Statements
-  const addReactImportStatements = () => {
-    testFileCode += `import { render, fireEvent } from '@testing-library/react'; 
-    import { build, fake } from 'test-data-bot'; 
-    import '@testing-library/jest-dom/extend-expect'
-    \n`;
-  };
-
-  // React Component Import Statement (Render Card)
-  const addComponentImportStatement = () => {
-    const componentPath = statements.componentPath;
-    let filePath = path.relative(projectFilePath, componentPath);
-    filePath = filePath.replace(/\\/g, '/');
-    testFileCode += `import ${statements.componentName} from '../${filePath}';`;
-  };
-
-  const addDescribeBlocks = () => {
-    describeBlocks.allIds.forEach((id) => {
-      testFileCode += `describe('${describeBlocks.byId[id].text}', () => {`;
-      addReactItStatement(id);
-      testFileCode += `}); \n`;
-    });
-  };
-
-  // React It Statements
-  const addReactItStatement = (describeId) => {
-    itStatements.allIds.forEach((itId) => {
-      if (itStatements.byId[itId].describeId === describeId) {
-        testFileCode += `it('${itStatements.byId[itId].text}', () => {`;
-        addReactStatements(itId);
-        testFileCode += '})';
-      }
-      testFileCode += '\n';
-    });
-  };
-
-  const addReactStatements = (itId) => {
-    const methods = identifyMethods(itId);
-    statements.allIds.forEach((id) => {
-      let statement = statements.byId[id];
-      if (statement.itId === itId) {
-        switch (statement.type) {
-          case 'action':
-            return addAction(statement);
-          case 'assertion':
-            return addAssertion(statement);
-          case 'render':
-            return addRender(statement, methods);
-          default:
-            return statement;
-        }
-      }
-    });
-  };
-
-  const identifyMethods = (itId) => {
-    const methods = new Set([]);
-    statements.allIds.forEach((id) => {
-      let statement = statements.byId[id];
-      if (statement.itId === itId) {
-        if (statement.type === 'action' || statement.type === 'assertion') {
-          methods.add(statement.queryVariant + statement.querySelector);
-        }
-      }
-    });
-    return Array.from(methods).join(', ');
-  };
-
-  // Render Jest Test Code
-  const addRender = (statement, methods) => {
-    let props = createRenderProps(statement.props);
-    testFileCode += `const {${methods}} = render(<${statements.componentName} ${props}/>);`;
-  };
-
-  // Render Props Jest Test Code
-  const createRenderProps = (props) => {
-    return props.reduce((acc, prop) => {
-      return acc + `${prop.propKey}={${prop.propValue}}`;
-    }, '');
-  };
-
-  // Action Jest Test Code
-  const addAction = (action) => {
-    if (action.eventValue) {
-      testFileCode += `fireEvent.${action.eventType}(${action.queryVariant + action.querySelector}
-                      ('${action.queryValue}'), { target: { value: ${action.eventValue} } });`;
-    } else {
-      testFileCode += `fireEvent.${action.eventType}(${action.queryVariant + action.querySelector}
-                      ('${action.queryValue}'));`;
-    }
-  };
-
-  // Assertion Jest Test Code
-  const addAssertion = (assertion) => {
-    testFileCode += `expect(${assertion.queryVariant + assertion.querySelector}
-      (${assertion.queryValue})).${assertion.matcherType}(${assertion.matcherValue});`;
-  };
+  // react has to have 2 states passed in
+  const generateTest = useGenerateTest('react', projectFilePath);
 
   const fileHandle = () => {
-    dispatchToGlobal(createFile(generatReactFile()));
+    dispatchToGlobal(createFile(generateTest(reactTestCase, mockData)));
     dispatchToGlobal(setFilePath(''));
   };
